+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 [![Donate via PayPal](https://img.shields.io/badge/Donate-PayPal-blue.svg?style=flat-square)](https://www.paypal.me/techtoday) 
 
 <p>
+<b>Version 1.0.36</b> November 2021<br/>
+- ImpulseUltimate: NEW: added <b>restart</b> command, to restart the sequence from scratch.</br>
+</p>
+<p>
 <b>Version 1.0.35</b> November 2021<br/>
 - Filterultimate and Invertultimate: added others checks to avoid processing invalid/non boolean convertible payloads and notify it in the node status.</br>
 </p>